pyproject.toml
[project]
name = "assort-health-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "JQ4386",email = "jingqi_qu@berkeley.edu"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "rich (>=14.2.0,<15.0.0)",
    "openai (>=2.8.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "requests (>=2.32.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)"
]

[tool.poetry]
packages = [{include = "assort_intake_bot"}]
